add a callback when component fully loads for figure, trafficLight and trafficSign
