+++ src/TrafficLight/TrafficLightSelector.types.ts
@@ -1,0 +1,6 @@
+import type { TrafficLightProps } from "./TrafficLight.types";
+
+export type TrafficLightSelectorProps = Omit<
+  TrafficLightProps,
+  "onTrafficLightLoad"
+> & { onImageLoad: (loaded: boolean) => void };
